Move preference related logic to ShadowsocksSettings
