+++ src/main/scala/com/github/shadowsocks/ShadowsocksSettings.scala
@@ -11,11 +11,61 @@ import android.os.{Build, Bundle}
 import android.preference.{Preference, PreferenceFragment, SwitchPreference}
 import android.support.v7.app.AlertDialog
 import android.webkit.{WebView, WebViewClient}
+import com.github.shadowsocks.database.Profile
+import com.github.shadowsocks.preferences.{DropDownPreference, NumberPickerPreference, PasswordEditTextPreference, SummaryEditTextPreference}
 import com.github.shadowsocks.utils.CloseUtils._
-import com.github.shadowsocks.utils.Key
+import com.github.shadowsocks.utils.{Key, Utils}
 
-// TODO: Move related logic here
+object ShadowsocksSettings {
+  // Constants
+  val PREFS_NAME = "Shadowsocks"
+  val PROXY_PREFS = Array(Key.profileName, Key.proxy, Key.remotePort, Key.localPort, Key.sitekey, Key.encMethod,
+    Key.isAuth)
+  val FEATURE_PREFS = Array(Key.route, Key.isProxyApps, Key.isUdpDns, Key.isIpv6)
+
+  // Helper functions
+  def updateDropDownPreference(pref: Preference, value: String) {
+    pref.asInstanceOf[DropDownPreference].setValue(value)
+  }
+
+  def updatePasswordEditTextPreference(pref: Preference, value: String) {
+    pref.setSummary(value)
+    pref.asInstanceOf[PasswordEditTextPreference].setText(value)
+  }
+
+  def updateNumberPickerPreference(pref: Preference, value: Int) {
+    pref.asInstanceOf[NumberPickerPreference].setValue(value)
+  }
+
+  def updateSummaryEditTextPreference(pref: Preference, value: String) {
+    pref.setSummary(value)
+    pref.asInstanceOf[SummaryEditTextPreference].setText(value)
+  }
+
+  def updateSwitchPreference(pref: Preference, value: Boolean) {
+    pref.asInstanceOf[SwitchPreference].setChecked(value)
+  }
+
+  def updatePreference(pref: Preference, name: String, profile: Profile) {
+    name match {
+      case Key.profileName => updateSummaryEditTextPreference(pref, profile.name)
+      case Key.proxy => updateSummaryEditTextPreference(pref, profile.host)
+      case Key.remotePort => updateNumberPickerPreference(pref, profile.remotePort)
+      case Key.localPort => updateNumberPickerPreference(pref, profile.localPort)
+      case Key.sitekey => updatePasswordEditTextPreference(pref, profile.password)
+      case Key.encMethod => updateDropDownPreference(pref, profile.method)
+      case Key.route => updateDropDownPreference(pref, profile.route)
+      case Key.isProxyApps => updateSwitchPreference(pref, profile.proxyApps)
+      case Key.isUdpDns => updateSwitchPreference(pref, profile.udpdns)
+      case Key.isAuth => updateSwitchPreference(pref, profile.auth)
+      case Key.isIpv6 => updateSwitchPreference(pref, profile.ipv6)
+    }
+  }
+}
+
 class ShadowsocksSettings extends PreferenceFragment with OnSharedPreferenceChangeListener {
+  import ShadowsocksSettings._
+
   private def activity = getActivity.asInstanceOf[Shadowsocks]
   lazy val natSwitch = findPreference(Key.isNAT).asInstanceOf[SwitchPreference]
   var stat: Preference = _
@@ -127,5 +177,29 @@ class ShadowsocksSettings extends PreferenceFragment with OnSharedPreferenceChan
       activity.attachService
     })
     case _ =>
+  }
+
+  def setEnabled(enabled: Boolean) {
+    findPreference(Key.isNAT).setEnabled(enabled)
+    for (name <- PROXY_PREFS) {
+      val pref = findPreference(name)
+      if (pref != null) pref.setEnabled(enabled)
+    }
+    for (name <- FEATURE_PREFS) {
+      val pref = findPreference(name)
+      if (pref != null) pref.setEnabled(enabled &&
+        (name != Key.isProxyApps || Utils.isLollipopOrAbove || !ShadowsocksApplication.isVpnEnabled))
+    }
+  }
+
+  def update(profile: Profile) {
+    for (name <- PROXY_PREFS) {
+      val pref = findPreference(name)
+      updatePreference(pref, name, profile)
+    }
+    for (name <- FEATURE_PREFS) {
+      val pref = findPreference(name)
+      updatePreference(pref, name, profile)
+    }
   }
 }
